test(cmd/git[run]) Cover Git.run(timeout=) end-to-end
why: The cmd-level timeout test mocks the underlying ``run()`` and
asserts the kwarg propagates -- proving plumbing but not behaviour.
The actual deadline path (real ``git`` subprocess, real selector
loop, real ``CommandTimeoutError``) was only covered at the
``_internal.run.run`` layer. Closing the loop with a direct
end-to-end test makes the public contract verifiable without
spelunking through the call graph.
what:
- Promote ``fast_timeout_constants`` from ``tests/_internal/test_run.py``
  to the root ``conftest.py`` so both ``tests/_internal`` and
  ``tests/cmd`` test suites can use it without duplication.
- Add ``test_git_run_timeout_e2e_raises_command_timeout`` exercising
  ``git cat-file --batch`` with ``stdin=subprocess.PIPE`` -- the child
  blocks reading stdin forever, so the deadline must fire and
  ``Git.run`` must raise ``CommandTimeoutError``.
- Wall-clock upper bound documented inline; uses the fast-constants
  fixture so the spread stays predictable on slow CI.
