+++ conftest.py
@@ -46,3 +46,21 @@ def setup(
     set_home: pathlib.Path,
 ) -> None:
     """Configure test fixtures for pytest."""
+
+
+@pytest.fixture
+def fast_timeout_constants(monkeypatch: pytest.MonkeyPatch) -> None:
+    """Tighten the deadline-loop spread so test wall-clock budgets stay reliable.
+
+    The two ``_TIMEOUT_*`` constants in :mod:`libvcs._internal.run` are tuned
+    for production use (0.5 s SIGTERM grace, 0.1 s selector poll). For tests
+    that intentionally fire the deadline, those defaults add up to ~0.7 s of
+    unavoidable wall-clock spread on top of the test's nominal timeout, which
+    makes upper-bound assertions fragile on loaded CI runners. This fixture
+    monkeypatches both to 0.05 s so the spread stays predictable; production
+    behaviour is unchanged.
+    """
+    from libvcs._internal import run as run_module
+
+    monkeypatch.setattr(run_module, "_TIMEOUT_KILL_GRACE_SECONDS", 0.05)
+    monkeypatch.setattr(run_module, "_TIMEOUT_POLL_INTERVAL_SECONDS", 0.05)
